refactor: getParticipated query 분리
