+++ web/src/main/java/org/cmc/curtaincall/web/party/PartyQueryController.java
@@ -1,21 +1,22 @@
 package org.cmc.curtaincall.web.party;
 
+import jakarta.validation.constraints.Size;
 import lombok.RequiredArgsConstructor;
 import org.cmc.curtaincall.domain.member.MemberId;
 import org.cmc.curtaincall.domain.party.PartyCategory;
 import org.cmc.curtaincall.domain.party.PartyId;
 import org.cmc.curtaincall.domain.party.dao.PartyDao;
 import org.cmc.curtaincall.domain.party.request.PartySearchParam;
-import org.cmc.curtaincall.domain.party.response.PartyDetailResponse;
-import org.cmc.curtaincall.domain.party.response.PartyParticipationResponse;
-import org.cmc.curtaincall.domain.party.response.PartyRecruitmentResponse;
-import org.cmc.curtaincall.domain.party.response.PartyResponse;
+import org.cmc.curtaincall.domain.party.response.*;
 import org.cmc.curtaincall.web.common.response.ListResult;
+import org.cmc.curtaincall.web.security.LoginMemberId;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.data.web.SortDefault;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
+
+import java.util.List;
 
 @RestController
 @RequiredArgsConstructor
@@ -55,5 +56,14 @@ public class PartyQueryController {
             @RequestParam(required = false) PartyCategory category, @PathVariable Long memberId
     ) {
         return new ListResult<>(partyDao.getParticipationList(pageable, new MemberId(memberId), category));
+    }
+
+    @GetMapping("/member/participated")
+    public ListResult<PartyParticipatedResponse> getParticipated(
+            @RequestParam @Validated @Size(max = 100) List<Long> partyIds, @LoginMemberId MemberId memberId
+    ) {
+        List<PartyParticipatedResponse> partyParticipatedResponses = partyDao.areParticipated(
+                partyIds.stream().map(PartyId::new).toList(), memberId);
+        return new ListResult<>(partyParticipatedResponses);
     }
 }
